Making pri patch to allow team_a branch to build for web again. Currently implmented is a dummy stub for ai diarazation to allow for web to build. This is most likely need to be imporved by the team to make an offical fix.
