+++ careconnect2025/frontend/lib/features/streaming_asr_with_diarization/streaming_asr_and_diarization_stub.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/material.dart';
+
+import '../../services/enhanced_file_service.dart';
+
+class StreamingAsrAndDiarizationScreen extends StatelessWidget {
+  final int? patientId;
+  final Function(FileUploadResponse)? onUploadSuccess;
+  final Function(String)? onUploadError;
+
+  const StreamingAsrAndDiarizationScreen({
+    super.key,
+    this.patientId,
+    this.onUploadSuccess,
+    this.onUploadError,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Padding(
+        padding: EdgeInsets.all(16),
+        child: Text(
+          'Speech-to-Text with Diarization is not available on Web.',
+          textAlign: TextAlign.center,
+        ),
+      ),
+    );
+  }
+}
